+++ src/core.rs
@@ -9,6 +9,7 @@ use self::flexible_gpus::FlexibleGpu;
 use self::load_balancers::LoadBalancer;
 use self::oos::Oos;
 use self::vms::Vm;
+use self::volumes::Volume;
 use self::vpn::Vpn;
 
 static HOURS_PER_MONTH: f32 = (365_f32 * 24_f32) / 12_f32;
@@ -17,6 +18,7 @@ pub mod flexible_gpus;
 pub mod load_balancers;
 pub mod oos;
 pub mod vms;
+pub mod volumes;
 pub mod vpn;
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -176,41 +178,6 @@ trait ResourceTrait {
     fn price_per_hour(&self) -> Result<f32, ResourceError>;
     fn compute(&mut self) -> Result<(), ResourceError>;
 }
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Volume {
-    pub osc_cost_version: Option<String>,
-    pub account_id: Option<String>,
-    pub read_date_rfc3339: Option<String>,
-    pub region: Option<String>,
-    pub resource_id: Option<String>,
-    pub price_per_hour: Option<f32>,
-    pub price_per_month: Option<f32>,
-    pub volume_type: Option<String>,
-    pub volume_size: Option<i32>,
-    pub volume_iops: Option<i32>,
-    pub price_gb_per_month: f32,
-    pub price_iops_per_month: f32,
-}
-
-impl ResourceTrait for Volume {
-    fn compute(&mut self) -> Result<(), ResourceError> {
-        let mut price_per_month = 0_f32;
-        price_per_month += (self.volume_size.unwrap() as f32) * self.price_gb_per_month;
-        price_per_month += (self.volume_iops.unwrap() as f32) * self.price_iops_per_month;
-        self.price_per_hour = Some(price_per_month / HOURS_PER_MONTH);
-        self.price_per_month = Some(price_per_month);
-        Ok(())
-    }
-
-    fn price_per_hour(&self) -> Result<f32, ResourceError> {
-        match self.price_per_hour {
-            Some(price) => Ok(price),
-            None => Err(ResourceError::NotComputed),
-        }
-    }
-}
-
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PublicIp {
     pub osc_cost_version: Option<String>,
